remove other cache args
